requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.7.2
    # via django
django==5.0.1
    # via
    #   -r requirements.in
    #   django-phonenumber-field
django-phonenumber-field[phonenumbers]==7.3.0
    # via -r requirements.in
phonenumbers==8.13.27
    # via django-phonenumber-field
pillow==10.2.0
    # via -r requirements.in
sqlparse==0.4.4
    # via django
tzdata==2023.4
    # via django
